backend/pyproject.toml
[project]
name = "openwhisper"
version = "0.1.0"
description = "A free and open source alternative to Whisperflow for Apple Silicon devices"
license='GPL-3.0-or-later'
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "faster-whisper>=1.2.1",
    "mlx-whisper>=0.4.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    "torch>=2.8.0",
    "uvicorn>=0.39.0",
]
